// src/components/organisms/SingleExchangeDetail/types.ts
export type IProps = {
    singleExchangeDetailsResultData: object;
  };
  export type SingleExchangeDetailProps = {
    name: string;
    country: string;
    trust_score_rank: number;
    image: string;
    year_established: number;
    description: string;
    facebook_url: string;
    reddit_url: string;
    telegram_url: string;
    slack_url: string;
    twitter_handle: string;
  };